Create README.md
This is an excerise introducing axion and how it does the get, put, post methods to the server. This example has couple of buttons which has event listeners which retrieves / updates user input. #11 is a post method which asks user to input data like  favorite food items via html and each time the user enters, it posts it to server and then prints it on the screen.when you enter many food item , each time, it will retrieve and display all of the previous ones with the new choices